+++ support/test/cljsbuild/test/compiler.clj
@@ -10,6 +10,10 @@
 
 (def cljs-path "src-cljs")
 (def crossover-path "crossovers")
+(def crossover-macro-absolute "/a/b/crossovers/macros.clj")
+(def crossover-macro-classpath "crossovers/macros.clj")
+(def crossover-macro-paths [{:absolute crossover-macro-absolute
+                             :classpath crossover-macro-classpath}])
 (def output-to "output-to")
 (def compiler-options
   {:output-to output-to
@@ -25,6 +29,7 @@
   (run-compiler
     cljs-path
     crossover-path
+    crossover-macro-paths
     compiler-options
     notify-command
     warn-on-undeclared?
@@ -32,9 +37,13 @@
     watch?) => nil
   (provided
     (fs/exists? output-to) => false :times 1
+    (util/find-files cljs-path #{"clj"}) => [] :times 1
     (util/find-files cljs-path #{"cljs"}) => ["src-cljs/a.cljs"] :times 1
     (util/find-files crossover-path #{"cljs"}) => ["crossovers/b.cljs"] :times 1
     (util/sh anything) => nil :times 1
     (fs/mod-time "src-cljs/a.cljs") => 1000 :times 1
+    (fs/mod-time "crossovers/b.cljs") => 1000 :times 1
+    (fs/mod-time crossover-macro-absolute) => 1000 :times 1
     (fs/mkdirs anything) => nil
+    (reload-clojure [crossover-macro-classpath] compiler-options) => nil :times 1
     (cljs/build cljs-path compiler-options) => nil :times 1))